backend/requirements.txt
# == API ==
Flask==3.1.2
flask-cors==6.0.2 # API ACCESS


# == DB ==
psycopg2-binary==2.9.11 # POSTGRESQL CONECTION
Flask-SQLAlchemy==3.1.1 # MODELS / ORM 
Flask-Migrate==4.1.0 # MIGRATIONS

# READ .ENV
python-dotenv